api/requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.4
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-rest-auth==2.1.4
Django==3.1.8
django-allauth==0.44.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-extensions==3.1.2
django-registration==3.1.2
django-sslserver==0.22
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
kombu==4.6.11
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
pytest==6.2.3
pytest-django==4.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
selenium==3.141.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
whitenoise==5.2.0
